backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23
Pillow>=9.0
scikit-learn>=1.2
umap-learn>=0.5.3
matplotlib>=3.5
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pandas>=2.0.0
python-multipart
sqlalchemy>=2.0.0